remove ingestion_job_state_provider + flatten inheritance hierarchy
