+++ metadata-ingestion/src/datahub/ingestion/api/ingestion_job_checkpointing_provider_base.py
@@ -1,31 +1,26 @@
+from abc import abstractmethod
 from dataclasses import dataclass
-from typing import Any, Dict, Type, TypeVar
+from typing import Any, Dict, List, NewType, Type, TypeVar
 
-from datahub.ingestion.api.committable import CommitPolicy
+import datahub.emitter.mce_builder as builder
+from datahub.configuration.common import ConfigModel
+from datahub.ingestion.api.committable import CommitPolicy, StatefulCommittable
 from datahub.ingestion.api.common import PipelineContext
-from datahub.ingestion.api.ingestion_job_state_provider import (
-    IngestionJobStateProvider,
-    IngestionJobStateProviderConfig,
-    JobId,
-    JobStateKey,
-    JobStatesMap,
-)
 from datahub.metadata.schema_classes import DatahubIngestionCheckpointClass
 
-#
-# Common type exports
-#
-JobId = JobId
-JobStateKey = JobStateKey
-
-#
-# Checkpoint state specific types
-#
+JobId = NewType("JobId", str)
 CheckpointJobStateType = DatahubIngestionCheckpointClass
-CheckpointJobStatesMap = JobStatesMap[CheckpointJobStateType]
+CheckpointJobStatesMap = Dict[JobId, CheckpointJobStateType]
 
 
-class IngestionCheckpointingProviderConfig(IngestionJobStateProviderConfig):
+@dataclass
+class JobStateKey:
+    pipeline_name: str
+    platform_instance_id: str
+    job_names: List[JobId]
+
+
+class IngestionCheckpointingProviderConfig(ConfigModel):
     pass
 
 
@@ -34,18 +29,17 @@ _Self = TypeVar("_Self", bound="IngestionCheckpointingProviderBase")
 
 @dataclass()
 class IngestionCheckpointingProviderBase(
-    IngestionJobStateProvider[CheckpointJobStateType]
+    StatefulCommittable[JobStateKey, CheckpointJobStatesMap]
 ):
     """
-    The base class(non-abstract) for all checkpointing state provider implementations.
-    This class is implemented this way as a concrete class is needed to work with the registry,
-    but we don't want to implement any of the functionality yet.
+    The base class for all checkpointing state provider implementations.
     """
 
     def __init__(
         self, name: str, commit_policy: CommitPolicy = CommitPolicy.ON_NO_ERRORS
     ):
-        super(IngestionCheckpointingProviderBase, self).__init__(name, commit_policy)
+        # Set the initial state to an empty dict.
+        super().__init__(name, commit_policy, {})
 
     @classmethod
     def create(
@@ -53,11 +47,27 @@ class IngestionCheckpointingProviderBase(
     ) -> "_Self":
         raise NotImplementedError("Sub-classes must override this method.")
 
+    @abstractmethod
     def get_last_state(
         self,
         state_key: JobStateKey,
     ) -> CheckpointJobStatesMap:
-        raise NotImplementedError("Sub-classes must override this method.")
+        ...
 
+    @abstractmethod
     def commit(self) -> None:
-        raise NotImplementedError("Sub-classes must override this method.")
+        ...
+
+    @staticmethod
+    def get_data_job_urn(
+        orchestrator: str,
+        pipeline_name: str,
+        job_name: JobId,
+        platform_instance_id: str,
+    ) -> str:
+        """
+        Standardizes datajob urn minting for all ingestion job state providers.
+        """
+        return builder.make_data_job_urn(
+            orchestrator, f"{pipeline_name}_{platform_instance_id}", job_name
+        )
