+++ metadata-ingestion/src/datahub/ingestion/source/state/redundant_run_skip_handler.py
@@ -3,7 +3,7 @@ from typing import Optional, cast
 
 import pydantic
 
-from datahub.ingestion.api.ingestion_job_state_provider import JobId
+from datahub.ingestion.api.ingestion_job_checkpointing_provider_base import JobId
 from datahub.ingestion.source.state.checkpoint import Checkpoint
 from datahub.ingestion.source.state.stateful_ingestion_base import (
     StatefulIngestionConfig,
